Initial pass at refactoring
This is the start of the work replacing the original design. In the
original we had nodes for each type of field that were responsible for
the validity, construction and general usage of an OpenAPI object.

This instead separates out the responsibility of validity and
construction into Factory classes for each node. These will also support
validating nodes without building them.