Rename project 😵
This renames the project from openapi_parser to openapi3_parser. This is
mostly to avoid confusion with the similarly named open_api_parser ruby
gem. Hopefully this also makes it easier for people to work out whether
this gem is for them as there is a lot of confusion as to whether
parsers of OpenAPI support v3 or not.

This could be a real pain if v4 is very similar to v3, but if it's
anything like v2 => v3 then we'd probably want to maintain a distinct
different gem. Either way that's a problem for future me.