New approach to input sources
This introduces 3 different methods to input into the parser: raw
(data), file or URL based.

There is a single source class which takes the input of a source_input.

This begins to set up the approach to cross file references with the
source reference class to model the data needed for a reference and the
reference resolver.