Build against a variety of Rubies
This tests against the current latest versions of Ruby 2.3 and 2.4 and JRuby.
It also allows fails on latest Ruby and JRuby.

I initially tried setting to 2.3 and 2.4 as aliases but this seemed to
cause problems as it was aliasing to a buggy version of Ruby 2.4 (2.4.1
which has a CGI unescape bug [1])

[1]: https://bugs.ruby-lang.org/issues/13539