Refactor Context object
This refactors the context object to be better suited to a world with
references in (they were previously somewhat cobbled together). It also
introduces two subclasses of Context: Location and Pointer.

Context::Pointer is used to represent a JSON pointer which indicates
where in a JSON hierarchy data is. Whereas Context::Location is used to
represent the shared information of a source file and a pointer.