Add a rudimentary versioning system
This is a little bit of a safety net for when there are new OpenAPI
versions that this library may not support.

The intention too is that in the future when there are other 3.x
versions we can do conditionals off the document class to check support.

I toyed with using a Kernel#warn initally for the warnings but it didn't
feel like stderr was always the place you wanted these so went for the
warnings attribute instead.