Email Validator
This wades into the firey pits of hell that is email validation - I've
wanted to be pretty leniant here and felt it cool to just use the one
W3C use in HTML spec.

It might even prove that it's better to even more leniant and just check
for an @ - we'll see.