Validatable class
This class is an attempt to solve the complexities involed in what to
pass around when validating nodes. Rather than passing around input or
context arguments this can be passed around which has access to those
fields and errors can be added directly to it rather than worrying about
returning strings or arrays.