Field node factory
This is a standalone class that represents a field for a node. I'm not
sure this should definitely exist as the only use of it is for
references and there isn't likely to be any more.

Though on the flipside it seems to be far too much logic to bung into a
reference field so for now the class stays.