CautiousDig class
This class has the purpose of acting like the dig method on arrays and
hashes but in a more cautious form where we don't care too much if the
types match.

It's imagined for use where we're taking a pointer string
"#/item/0/blah" where we have a string and then need to look up a hash,
then an array and then a hash and we wouldn't know from that string
which type each item should be.