Handle recursive references better
This is a fix for https://github.com/kevindew/openapi3_parser/issues/4

This library populates the data of nodes and does validation by working
out the entire contents and inputting that. This presents us with a
significant problem with recursive data as we can never know the full
input - and is really something I should have thought more about sooner.

The approach I've taken to solving this is to use decorator objects
which point to a location in the document and can terminate recursion.
These objects can then be used at runtime to traverse infinitely (memory
permitting) down a recursive tree of nodes.

The modifications for this are:

Allow context to look up the input/node for itself by asking the
document what is at that location. This allows a lazy approach to
accessing the data. This has meant that the context needs to know if it
is itself a reference as the input/node for this is actually at the
parent item.

When we ask a reference field for it's resolved input this can now
return a hash of data (as before) or a RecursiveResolvedInput object
which is hash like but accesses data from the root document.

When we try to build a node on a Reference node we can now receive an
exception when it's in a recursive loop (as that can never build) so
instead we have a method that returns a RecursivePointer object that we
can use the node method on to access the node. This means we don't look
up the actual node until it's needed and at that point we can access it
via the Document.

All of this works on the basis we don't need to access the recursive
data until the document itself is built. We could still get caught in
loops if this data is accessed prematurely and a future optimisation
could be to guard against this if this becomes a regular problem.

Hopefully I've not missed any other scenarios that can fall into a loop.