Initialise DSL variables when we access them
Accessing these variables causes Ruby warnings when they're not set.
It doesn't seem like it's trivial to initialise them in advance (can't
use the extended method) so simplest thing seems to be to use the ||=
operator whenever we access them since we don't know if they're defined
or not.

This seemed less verbose than using instance_variable_defined? on each
of them.