Handle an unresolvable recursive reference
This makes a number of changes to the reference system so that recursive
references that never resolve (due to infinite references) don't crash
this parser.

Rather than storing just the single location that referenced a node both
the node factory and node context objects now store an array of
reference locations, this allows traversing the journey back to the
original reference.

The ReferenceRegistry has been modified to be something where factories are
registered too and then later accessed back. This behaviour means that
it can act the same for recursive and non-recursive references, where
previously there was a somewhat odd build argument passed to determine
between these.

In this change it has become easier to re-merge the Field::RecursiveReference
and Field::Reference classes as they now behave in a more similar
fashion.

To handle the specific case of an eternal loop of references an
additional validation method was added to a Field::Reference which
recursively loops through each resolved reference until it either finds
the original reference again - indicating it will never complete - or a
non-reference node.

In putting this together I experimented with a class that represented an
unresolved reference which was a form of null object pattern for
Source::ResolvedReference. However I decided this made life too complex
for a Field::Reference class which already dealt with an invalid
reference field.