Raise error if resolved_reference is nil
This fixes a todo loose end where there was a mixture of code using and
not using the safe navigation operator on the same variable. As far as I
can tell this method shouldn't be executed unless the variable is not
nil, but since that is dependant on code outside this method it seems
reasonable to have an exception raised were the error scenario to occur.