Test to check for no duplicate reference errors
This adds a test that provides a layer of protection against reference
errors being pointed out twice in a document validation. It's an area
that I figured we could easily accidentally start spitting out duplicate
errors so I tried to think of a good scenario to test it with.

This came out a bit convoluted and maybe it might be suited better as an
integration test?