Add missing require forwardable
These files were inconsistent with the rest of the app that had specific
requires for when the forwardable module was included. I debated about
whether to move all the requires to the gem entry file but decided
against it as was nice to have usages in context (e.g. commonmarker just
for the markdown file) since Ruby does no enforcement this might be
ultimately pointless though.