Fix bug on each iterators for Node::Object and Node::Map
This resolves an issue where these were incorrectly treating the data as
an array to be iterated through whereas previously this was a hash.

To resolve this I've moved this code to a single place, the Placeholder
class, which can do the iteration and resolving directly. I've also
added an integration test for this functionality the reduce the risk of
this happening again.