Method to lookup schema name
This provides a method that can lookup a schema name based on it's
source location.

For example given an OpenAPI document of:

```
components:
  schemas:
    Pet:
      type: object
```

The schema name for this object is Pet.

Now this sort of naming look up is really unusual (and to me seems
rather inconsistent) in the scope of OpenAPI. As far as I'm aware this
is the only place where a field name in a components object can have a
role - and this makes things somewhat confusing if the item is
referenced to an external file. Schema name is a concept in
the [OpenAPI Specification][oas] so it seems to be a first class thing.
However life would be a lot simpler if the title field was used instead
of this as a convention.

This uses the position in the source to decide whether this is a named
schema or not. This decision is made if this item is in a path where the
item two above it is is named schemas (as is the case
with #/components/schemas/Pet for instance) otherwise it is not
considered a named schema.

[oas]: https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.2.md#composition-and-inheritance-polymorphism