Add equal methods to Node::Context
This adds methods that check whether two Node::Contexts are the same
(same document location, same source location and same input) and a more
forgiving equality check to see if just the source_location and input
match.

The intention is that the second method can be used in nodes themselves
to check equality between two referenced nodes.