Add equality checks for nodes
This allows checking whether a node is the same as another. By the same
it refers to the same type at the same location. This means that if a
node is referenced each referenced version of it will return true to
`==` despite actually sitting at different document locations.

I anticipate this to be a useful method for `.uniq` on collections of
schemas.