Use cascading servers on Operation objects
This sets operation objects to use the servers data of the parent object
(Path Item) when servers aren't explicitly defined on an Operation
object. As a Path Item may also have servers that cascade from the root
object the servers on an operation could also be from the root object.

This is to meet the Operation specification:
https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.2.md#fixed-fields-8