Resolve Psych warnings
The interface for Psych changed in version 3.1 and emits warnings when
the legacy interface is used. This is a somewhat challenging situation
given that this is a gem that is part of Ruby's standard library and
Ruby versions before 2.6 don't have the new interface.

To resolve this this gem now has an explicit dependency on Psych
starting from the version where this new interface was introduced.