Enhance recursion integration test to cover more cases
In https://github.com/kevindew/openapi3_parser/issues/13 a problem was
pointed out where we fell into a recursive loop when accessing a
recursive item that was a property of a map.

Since these things are total pains to test in unit tests (so much
set-up) I've set up a variety of integration tests to cover a few cases.