Don't return errors on a reference in a recursive loop
Once a node has referenced itself we can fall into a recursive loop
where it keeps trying to check it's child nodes (itself) for errors.
There isn't any value in looking for errors once we're in a recursive
loop as we have effectively already visited and validated the item in
question.

This is added in to prevent arrays and maps from being at risk from
falling into a recursive loop. We introduced a fix for objects back
in be5fa545, however this didn't cover all cases.