Remove explicit Psych dependency
This resolves an issue where this gem would affect the version of Psych
that is installed in applications. This becomes problematic with Ruby
3.1 where Psych 4 is part of the stdlib and this gem will prevent using
that.

From my research it's very rare for an app to specify a particular psych
dependency and instead it's more common to rely on Ruby stdlib to
provide a compatible one - which should be the case for all the Ruby
versions this gem operates with.