Add OpenapiVersion class for version comparisons
This class is to replace the use of a string as the means for
representing the version of an OpenAPI version. A new class has been
added to make it simpler to compare versions in a more sophisticated way
than strings. The class inherits from Gem::Version which is a part of
Ruby stdlib which has this comparison logic built in.

The spaceship operator method has been replaced with one that will
apply polymorphish to comparisons. E.g:

It replaces the need for:

```
OpenapiVersion.new("3.11") > OpenapiVersion.new("3.2")
```

with:

```
OpenapiVersion.new("3.11") > "3.2"
```

I added this in as it felt more natural to do this than have some
contrived `equal_or_greater_than?` methods. However I do have a concern
that I may have created some Ruby magic