Version specific allow_extensions for Discriminator
A change in OpenAPI 3.1 is that the Discriminator object can accept
extensions, when previously this was not allowed.

In order to code this I've had to adjust the DSL so that the
allow_extensions method can accept a block.

We didn't have unit tests for the DSL so I'm only testing this at the
node level. This should be ok as it has full coverage.