Share behaviour between schema objects
This sets up the ability for shared code between Schema objects and
providing customisation for ones of different OpenAPI versions.

It marks a direction towards modelling the specification of JSON Schema
2020-12 JSON Schema Validation [1] specifically rather trying to
understand how different dialects could be used. My expectation is that
it is very much an edge case that someone would use another dialect and
it seems incredibly hard to consider how they can be supported.

There's likely some optimising that could be done of the shared example
specs, but I wanted to get something together on the faster side.

This also has a rename of Schema::OasDialect3_1 to Schema::V3_1.
I don't think we'll ever try to model other dialects so I think it's
best to have a single class to try model the schemas for 3.1 and
hopefully later versions.

[1]: https://json-schema.org/draft/2020-12/draft-bhutton-json-schema-validation-01

WIP