Configure type field for OpenAPI 3.1
This field is rather complicated as JSON Schema 2020-12 allows it to be
an array or a string and other parts of OpenAPI don't allow different
types.

Another difference is that JSON Schema specification doesn't have the
rule (or I didn't find it) where you require items to be specified if
type is array. There are different types of items (such as prefixItems
and additionalItems) in JSON Schema so I guess the items rule is
complex.