Correct handling of minimum and maximum of schema
This fixes an error in OpenAPI 3.0 schema handling where minimum and
maximum were flagged as integers when actually they are numeric.

It then corrects the configuration of exclusiveMaximum and
exclusiveMinimum for OpenAPI 3.1 where JSON Schema 2021 [1] defines
these as numeric values compared to the boolean values they held in
OpenAPI 3.0 (via JSON Schema 2017 [2]).

[1]: https://datatracker.ietf.org/doc/html/draft-bhutton-json-schema-validation-00#section-6.2.3
[2]: https://datatracker.ietf.org/doc/html/draft-wright-json-schema-validation-00#section-5.3