Add patternProperties to Schema
Based on JSON schema 2021 [1]. Where the value is expected to be a map
where the key is a regex (to match a string property name) and value is
a Schema object.

Like the implementation of pattern in this codebase this doesn't have
validation that pattern is a regex, which could be added.

[1]: https://datatracker.ietf.org/doc/html/draft-bhutton-json-schema-00#section-10.3.1.1