Support JsonSchemas that are a boolean value
JsonSchema 2020-12 has a rather complex quirk that providing a boolean
is treated as a valid schema. Where a value of `true` means the schema
will pass anything and a value of `false` means the schema will pass
nothing [1].

This is quite a pain to model as there is an assumption in this
libraries code that a node can only be one type and that all references
resolve to a map type.

This code configures the factory to produce a v3_1 schema node when
given these, in a subsequent commit I will configure the node class to
make use of these.

[1]: https://datatracker.ietf.org/doc/html/draft-bhutton-json-schema-00#section-4.3.2