Add some dubious hacks to get appropriate errors
These build on the smell of hacks from the previous commit to make this
class a bit smelly.

This adds some work arounds to get appropriate error messages for nodes
that could be one of two types. Should we need more usage of this we
probably want to refactor aspects of the ObjectFactory code to be able
to handle dual types and the TypeChecker.