Add output of warnings to Kernel#warn
This sends warnings for a document to Kernel#warn so they are output for
a user. A document can be created with emit_warnings: false as a means
to suppress these.

Previously we only used warnings as an attribute of a document which
could be accessed which meant they were quite hidden. This was based on
the expectation that users are opening schemas they don't know about and
thus a warning isn't helpful. However I now think for most people they
are working with a schema in their codebase and thus it is better to be
more explicit.