JSON schema dialect warnings
This adds a system in place to emit warnings if a schema is parsed and
it doesn't have the OAS schema dialect for OpenAPI 3.1.

My expectation is very few people will consider using over schema
dialects and likely no-one will need this message, but if you did it
could be quite confusing if this gem just seems to perform oddly with a
different schema dialect.

As this was relatively uncharted territory for this gem I've had to be a
bit creative with implementing this. Without a clear place to hook this
in, I've added it to the validation route of schemas. Since schemas are
lazily parsed, I've made warnings for a document be a lazy loaded
attribute that forces a validation run first so it can provide a
complete list.

I didn't want this gem to be super annoying and output warnings on every
schema so I've configured this to only warn once per unsupported schema
dialect.